+++ blog/home/HomePage/Content/index.js
@@ -1,0 +1,11 @@
+import React from 'react'
+import { Layout } from 'antd'
+const { Content } = Layout
+
+const ContentLayout = () => (
+    <Content className="">
+        我是content
+    </Content>
+)
+
+export default ContentLayout