+++ blog/src/index.js
@@ -16,5 +16,5 @@ ReactDOM.render(
 )
 
 if (module.hot) {
-  module.hot.accept();
+  module.hot.accept()
 }